feat: :sparkles: Refactor serializers into separate modules for better organization and maintainability
