+++ app/lotus/serializers/locais.py
@@ -1,0 +1,27 @@
+from typing import ClassVar
+
+from rest_framework import serializers
+
+from lotus.models import Bloco, Sala
+
+
+class BlocoSerializer(serializers.ModelSerializer):
+    """Serializer de blocos."""
+
+    class Meta:
+        """Meta informações do serializer."""
+
+        model = Bloco
+        fields = "__all__"
+
+
+class SalaSerializer(serializers.ModelSerializer):
+    """Serializer de salas."""
+
+    bloco = BlocoSerializer()
+
+    class Meta:
+        """Meta informações do serializer."""
+
+        model = Sala
+        fields: ClassVar[list[str]] = ["id", "nome", "bloco"]
